PR feedback
Move DisableBuildOnlyContainerValidation to the Pipeline.
